pantheon.mutter43: Remove keybinding support for window shading
This fixes a crash when gsettings-desktop-schemas >= 45.alpha is used: Settings schema
'org.gnome.desktop.wm.keybindings' does not contain a key named 'toggle-shaded'.

We should see if it is possible to fetch the fix from magpie when this is fixed there.
It is always dangerous to apply a MR partially :-(

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2884